Add fix for repeating text & add better prints
"The Match Begins" was being repeatedly printed because the status
sometimes equaled "2" (this might be when the match ended, but payouts
haven't been distributed). Now we only look for the status not being
locked".

Also add the diagnostic prints used to debug this.

Also add a template dictionary for what we will be sending to the
database.