Remove redundancy in match.py, bet.py
Added an enum in bet.py to replace the need for a function for player1
and player2. On top of this, the print statements for debugging in
bet.py were replaced by an assert. This makes the code much nicer on
the eyes.

Match.py had it's record_match function deleted as I have decided that
the main.py runloop is the proper place for that. It is TBH how I will
use match information now.